TINY-14123: add request filter and request matcher
