+++ modules/agar/src/main/ts/ephox/agar/api/FetchSpy.ts
@@ -1,2 +1,2 @@
 export type { FetchSpyOptions } from '../http/FetchSpy';
-export { matchGet, matchPost, matchPut, matchDel, matchPatch, pWithFetchSpy } from '../http/FetchSpy';
+export { pWithFetchSpy } from '../http/FetchSpy';
