+++ modules/agar/src/main/ts/ephox/agar/api/RequestFilter.ts
@@ -1,0 +1,1 @@
+export { makeRequestFilter, get, post, put, del, patch } from '../http/RequestFilter';
