+++ modules/agar/src/main/ts/ephox/agar/api/RequestMatcher.ts
@@ -1,0 +1,1 @@
+export { makeRequestMatcher, get, post, put, del, patch } from '../http/RequestMatcher';
